refactor: separate button variants for fast refresh
